pyproject.toml
[tool.poetry]
name = "librarymanagement"
version = "0.1.0"
description = "library user management api"
authors = ["Sujan Midatani <sujanmidatani7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
pydantic = "^2.6.4"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
pymongo = {extras = ["srv"], version = "^4.6.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
